// frontend/src/types/Atcions.typs.ts
export enum ActionTypes {
  GET_CURRENT_USER = "GET_CURRENT_USER",
  GET_CURRENT_USER_APPPOINTMENTS = "GET_CURRENT_USER_APPPOINTMENTS",
  GET_ALL_USERS = "GET_ALL_USERS",
  GET_USER_BY_ID = "GET_USER_BY_ID",
  CREATE_USER = "CREATE_USER",
  UPDATE_USER = "UPDATE_USER",
  DELETE_USER = "DELETE_USER",
  GET_DOCTOR_AVILABILITY = "GET_DOCTOR_AVILABILITY",
  LOGIN = "LOGIN",
  REGISTER = "REGISTER",
  LOGOUT = "LOGOUT",
  GET_USER_APPOINTMENTS = "GET_USER_APPOINTMENTS",
  GET_USER_APPOINTMENTS_BY_ID = "GET_USER_APPOINTMENTS_BY_ID",
  CREATE_APPOINTMENT = "CREATE_APPOINTMENT",
  UPDATE_APPOINTMENT = "UPDATE_APPOINTMENT",
  DELETE_APPOINTMENT = "DELETE_APPOINTMENT",
}